+++ src/layouts/components/LayoutSideNav.vue
@@ -1,0 +1,37 @@
+<template>
+  <l-side-nav
+    v-if="settingStore.showSidebar"
+    :show-logo="settingStore.showSidebarLogo"
+    :layout="settingStore.layout"
+    :is-fixed="settingStore.isSidebarFixed"
+    :menu="sideMenu"
+    :theme="settingStore.mode"
+    :is-compact="settingStore.isSidebarCompact"
+  />
+</template>
+
+<script setup lang="ts">
+import { computed } from 'vue';
+import { useRoute } from 'vue-router';
+import { storeToRefs } from 'pinia';
+import { usePermissionStore, useSettingStore } from '@/store';
+import LSideNav from './SideNav.vue';
+
+const route = useRoute();
+const permissionStore = usePermissionStore();
+const settingStore = useSettingStore();
+const { routers: menuRouters } = storeToRefs(permissionStore);
+
+const sideMenu = computed(() => {
+  const { layout, splitMenu } = settingStore;
+  let newMenuRouters = menuRouters.value;
+  if (layout === 'mix' && splitMenu) {
+    newMenuRouters.forEach((menu) => {
+      if (route.path.indexOf(menu.path) === 0) {
+        newMenuRouters = menu.children.map((subMenu) => ({ ...subMenu, path: `${menu.path}/${subMenu.path}` }));
+      }
+    });
+  }
+  return newMenuRouters;
+});
+</script>
