+++ src/layouts/index.vue
@@ -1,0 +1,93 @@
+<template>
+  <div>
+    <template v-if="setting.layout.value === 'side'">
+      <t-layout key="side" :class="mainLayoutCls">
+        <t-aside><layout-side-nav /></t-aside>
+        <t-layout>
+          <t-header><layout-header /></t-header>
+          <t-content><layout-content /></t-content>
+        </t-layout>
+      </t-layout>
+    </template>
+
+    <template v-else>
+      <t-layout key="no-side">
+        <t-header />
+        <t-layout :class="mainLayoutCls">
+          <layout-side-nav />
+          <layout-content />
+        </t-layout>
+      </t-layout>
+    </template>
+    <setting-com />
+  </div>
+</template>
+
+<script setup lang="ts">
+import { computed, onMounted, watch, onBeforeUnmount } from 'vue';
+import { storeToRefs } from 'pinia';
+import { useRoute } from 'vue-router';
+import { useSettingStore, useTabsRouterStore } from '@/store';
+
+import SettingCom from './setting.vue';
+import LayoutHeader from './components/LayoutHeader.vue';
+import LayoutContent from './components/LayoutContent.vue';
+import LayoutSideNav from './components/LayoutSideNav.vue';
+
+import { prefix } from '@/config/global';
+
+import '@/style/layout.less';
+
+const route = useRoute();
+const settingStore = useSettingStore();
+const tabsRouterStore = useTabsRouterStore();
+const setting = storeToRefs(settingStore);
+
+const mainLayoutCls = computed(() => [
+  {
+    't-layout--with-sider': settingStore.showSidebar,
+  },
+]);
+
+const appendNewRoute = () => {
+  const {
+    path,
+    query,
+    meta: { title },
+    name,
+  } = route;
+  tabsRouterStore.appendTabRouterList({ path, query, title: title as string, name, isAlive: true });
+};
+
+const getTabRouterListCache = () => {
+  tabsRouterStore.initTabRouterList(JSON.parse(localStorage.getItem('tabRouterList')));
+};
+const setTabRouterListCache = () => {
+  const { tabRouters } = tabsRouterStore;
+  localStorage.setItem('tabRouterList', JSON.stringify(tabRouters));
+};
+
+onMounted(() => {
+  appendNewRoute();
+});
+
+// 如果不需要持久化标签页可以注释掉以下的 onMounted 和 onBeforeUnmount 的内容
+onMounted(() => {
+  if (localStorage.getItem('tabRouterList')) getTabRouterListCache();
+  window.addEventListener('beforeunload', setTabRouterListCache);
+});
+
+onBeforeUnmount(() => {
+  window.removeEventListener('beforeunload', setTabRouterListCache);
+});
+
+watch(
+  () => route.path,
+  () => {
+    appendNewRoute();
+    document.querySelector(`.${prefix}-layout`).scrollTo({ top: 0, behavior: 'smooth' });
+  },
+);
+</script>
+
+<style lang="less" scoped></style>
