+++ src/router/modules/components.ts
@@ -1,4 +1,4 @@
-import Layout from '@/layouts';
+import Layout from '@/layouts/index.vue';
 import ListIcon from '@/assets/assets-slide-list.svg';
 import FormIcon from '@/assets/assets-slide-form.svg';
 import DetailIcon from '@/assets/assets-slide-detail.svg';
